+++ packages/Modal/index.js
@@ -1,4 +1,5 @@
 import Modal from './Modal.component';
+import WizardModal from './WizardModal.component';
 import GenericModal from './GenericModal.component';
 
-export { Modal, GenericModal };
+export { Modal, GenericModal, WizardModal };