+++ packages/Disclaimer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.10](https://github.com/comparethemarketau/manor-react/compare/@comparethemarketau/manor-disclaimer@0.14.9...@comparethemarketau/manor-disclaimer@0.14.10) (2021-08-17)
+
+**Note:** Version bump only for package @comparethemarketau/manor-disclaimer
+
+
+
+
+
 ## [0.14.9](https://github.com/comparethemarketau/manor-react/compare/@comparethemarketau/manor-disclaimer@0.14.8...@comparethemarketau/manor-disclaimer@0.14.9) (2021-07-29)
 
 **Note:** Version bump only for package @comparethemarketau/manor-disclaimer